Rename dynamic_types to dynamic_types_collection
