+++ ddsrecorder_participants/src/cpp/common/types/dynamic_types_collection/v1/DynamicTypesCollection.cpp
@@ -1,4 +1,4 @@
-// Copyright 2016 Proyectos y Sistemas de Mantenimiento SL (eProsima).
+// Copyright 2023 Proyectos y Sistemas de Mantenimiento SL (eProsima).
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -14,9 +14,9 @@
 
 /*!
  * @file DynamicTypesCollection.cpp
- * This source file contains the implementation of the described types in the IDL file.
+ * This source file contains the definition of the described types in the IDL file.
  *
- * This file was generated by the tool fastddsgen.
+ * This file was generated by the tool gen.
  */
 
 #ifdef _WIN32
@@ -26,29 +26,34 @@ char dummy;
 }  // namespace
 #endif  // _WIN32
 
-#include <ddsrecorder_participants/common/types/dynamic_types/v2/DynamicTypesCollection.hpp>
-
-#if FASTCDR_VERSION_MAJOR > 1
-
+#include <ddsrecorder_participants/common/types/dynamic_types_collection/v1/DynamicTypesCollection.hpp>
 #include <fastcdr/Cdr.h>
 
-
 #include <fastcdr/exceptions/BadParamException.h>
-using namespace eprosima::fastcdr::exception;
 
 #include <utility>
 
-
 namespace eprosima {
-
 namespace ddsrecorder {
-
 namespace participants {
 
+using namespace eprosima::fastcdr::exception;
 
+#define DynamicTypesCollection_max_cdr_typesize 78004ULL;
+#define DynamicType_max_cdr_typesize 780ULL;
+#define DynamicTypesCollection_max_key_cdr_typesize 0ULL;
+#define DynamicType_max_key_cdr_typesize 0ULL;
 
 DynamicType::DynamicType()
 {
+    // string m_type_name
+    m_type_name = "";
+
+    // string m_type_information
+    m_type_information = "";
+
+    // string m_type_object
+    m_type_object = "";
 }
 
 DynamicType::~DynamicType()
@@ -74,28 +79,27 @@ DynamicType::DynamicType(
 DynamicType& DynamicType::operator =(
         const DynamicType& x)
 {
-
     m_type_name = x.m_type_name;
     m_type_information = x.m_type_information;
     m_type_object = x.m_type_object;
+
     return *this;
 }
 
 DynamicType& DynamicType::operator =(
         DynamicType&& x) noexcept
 {
-
     m_type_name = std::move(x.m_type_name);
     m_type_information = std::move(x.m_type_information);
     m_type_object = std::move(x.m_type_object);
+
     return *this;
 }
 
 bool DynamicType::operator ==(
         const DynamicType& x) const
 {
-    return (m_type_name == x.m_type_name &&
-           m_type_information == x.m_type_information &&
+    return (m_type_name == x.m_type_name && m_type_information == x.m_type_information &&
            m_type_object == x.m_type_object);
 }
 
@@ -103,6 +107,42 @@ bool DynamicType::operator !=(
         const DynamicType& x) const
 {
     return !(*this == x);
+}
+
+size_t DynamicType::getMaxCdrSerializedSize(
+        size_t current_alignment)
+{
+    static_cast<void>(current_alignment);
+    return DynamicType_max_cdr_typesize;
+}
+
+size_t DynamicType::getCdrSerializedSize(
+        const DynamicType& data,
+        size_t current_alignment)
+{
+    static_cast<void>(data);
+    size_t initial_alignment = current_alignment;
+
+    current_alignment += 4 + eprosima::fastcdr::Cdr::alignment(current_alignment, 4) + data.type_name().size() + 1;
+    current_alignment += 4 +
+            eprosima::fastcdr::Cdr::alignment(current_alignment, 4) + data.type_information().size() + 1;
+    current_alignment += 4 + eprosima::fastcdr::Cdr::alignment(current_alignment, 4) + data.type_object().size() + 1;
+
+    return current_alignment - initial_alignment;
+}
+
+void DynamicType::serialize(
+        eprosima::fastcdr::Cdr& scdr) const
+{
+    scdr << m_type_name.c_str(); scdr << m_type_information.c_str(); scdr << m_type_object.c_str();
+}
+
+void DynamicType::deserialize(
+        eprosima::fastcdr::Cdr& dcdr)
+{
+    dcdr >> m_type_name;
+    dcdr >> m_type_information;
+    dcdr >> m_type_object;
 }
 
 /*!
@@ -219,6 +259,24 @@ std::string& DynamicType::type_object()
     return m_type_object;
 }
 
+size_t DynamicType::getKeyMaxCdrSerializedSize(
+        size_t current_alignment)
+{
+    static_cast<void>(current_alignment);
+    return DynamicType_max_key_cdr_typesize;
+}
+
+bool DynamicType::isKeyDefined()
+{
+    return false;
+}
+
+void DynamicType::serializeKey(
+        eprosima::fastcdr::Cdr& scdr) const
+{
+    static_cast<void>(scdr);
+}
+
 DynamicTypesCollection::DynamicTypesCollection()
 {
 }
@@ -242,16 +300,16 @@ DynamicTypesCollection::DynamicTypesCollection(
 DynamicTypesCollection& DynamicTypesCollection::operator =(
         const DynamicTypesCollection& x)
 {
-
     m_dynamic_types = x.m_dynamic_types;
+
     return *this;
 }
 
 DynamicTypesCollection& DynamicTypesCollection::operator =(
         DynamicTypesCollection&& x) noexcept
 {
-
     m_dynamic_types = std::move(x.m_dynamic_types);
+
     return *this;
 }
 
@@ -267,12 +325,48 @@ bool DynamicTypesCollection::operator !=(
     return !(*this == x);
 }
 
+size_t DynamicTypesCollection::getMaxCdrSerializedSize(
+        size_t current_alignment)
+{
+    static_cast<void>(current_alignment);
+    return DynamicTypesCollection_max_cdr_typesize;
+}
+
+size_t DynamicTypesCollection::getCdrSerializedSize(
+        const DynamicTypesCollection& data,
+        size_t current_alignment)
+{
+    static_cast<void>(data);
+    size_t initial_alignment = current_alignment;
+
+    current_alignment += 4 + eprosima::fastcdr::Cdr::alignment(current_alignment, 4);
+
+    for (size_t a = 0; a < data.dynamic_types().size(); ++a)
+    {
+        current_alignment += DynamicType::getCdrSerializedSize(data.dynamic_types().at(a), current_alignment);
+    }
+
+    return current_alignment - initial_alignment;
+}
+
+void DynamicTypesCollection::serialize(
+        eprosima::fastcdr::Cdr& scdr) const
+{
+    scdr << m_dynamic_types;
+}
+
+void DynamicTypesCollection::deserialize(
+        eprosima::fastcdr::Cdr& dcdr)
+{
+    dcdr >> m_dynamic_types;
+}
+
 /*!
  * @brief This function copies the value in member dynamic_types
  * @param _dynamic_types New value to be copied in member dynamic_types
  */
 void DynamicTypesCollection::dynamic_types(
-        const std::vector<eprosima::ddsrecorder::participants::DynamicType>& _dynamic_types)
+        const std::vector<DynamicType>& _dynamic_types)
 {
     m_dynamic_types = _dynamic_types;
 }
@@ -282,7 +376,7 @@ void DynamicTypesCollection::dynamic_types(
  * @param _dynamic_types New value to be moved in member dynamic_types
  */
 void DynamicTypesCollection::dynamic_types(
-        std::vector<eprosima::ddsrecorder::participants::DynamicType>&& _dynamic_types)
+        std::vector<DynamicType>&& _dynamic_types)
 {
     m_dynamic_types = std::move(_dynamic_types);
 }
@@ -291,7 +385,7 @@ void DynamicTypesCollection::dynamic_types(
  * @brief This function returns a constant reference to member dynamic_types
  * @return Constant reference to member dynamic_types
  */
-const std::vector<eprosima::ddsrecorder::participants::DynamicType>& DynamicTypesCollection::dynamic_types() const
+const std::vector<DynamicType>& DynamicTypesCollection::dynamic_types() const
 {
     return m_dynamic_types;
 }
@@ -300,19 +394,29 @@ const std::vector<eprosima::ddsrecorder::participants::DynamicType>& DynamicType
  * @brief This function returns a reference to member dynamic_types
  * @return Reference to member dynamic_types
  */
-std::vector<eprosima::ddsrecorder::participants::DynamicType>& DynamicTypesCollection::dynamic_types()
+std::vector<DynamicType>& DynamicTypesCollection::dynamic_types()
 {
     return m_dynamic_types;
 }
 
-} // namespace participants
+size_t DynamicTypesCollection::getKeyMaxCdrSerializedSize(
+        size_t current_alignment)
+{
+    static_cast<void>(current_alignment);
+    return DynamicTypesCollection_max_key_cdr_typesize;
+}
 
+bool DynamicTypesCollection::isKeyDefined()
+{
+    return false;
+}
 
-} // namespace ddsrecorder
+void DynamicTypesCollection::serializeKey(
+        eprosima::fastcdr::Cdr& scdr) const
+{
+    static_cast<void>(scdr);
+}
 
-
-} // namespace eprosima
-// Include auxiliary functions like for serializing/deserializing.
-#include "DynamicTypesCollectionCdrAux.ipp"
-
-#endif // FASTCDR_VERSION_MAJOR > 1
+} /* namespace participants */
+} /* namespace ddsrecorder */
+} /* namespace eprosima */
