+++ ddsrecorder_participants/src/cpp/common/types/dynamic_types_collection/v2/DynamicTypesCollectionCdrAux.ipp
