+++ ddsrecorder_participants/src/cpp/common/types/dynamic_types_collection/v2/DynamicTypesCollectionPubSubTypes.cpp
@@ -22,7 +22,7 @@
 
 #include <fastdds/rtps/common/CdrSerialization.hpp>
 
-#include <ddsrecorder_participants/common/types/dynamic_types/v2/DynamicTypesCollectionPubSubTypes.hpp>
+#include <ddsrecorder_participants/common/types/dynamic_types_collection/v2/DynamicTypesCollectionPubSubTypes.hpp>
 #include "DynamicTypesCollectionCdrAux.hpp"
 
 using SerializedPayload_t = eprosima::fastrtps::rtps::SerializedPayload_t;
