+++ ddsrecorder_participants/src/cpp/common/types/dynamic_types_collection/v2/DynamicTypesCollectionv1.cpp
@@ -1,4 +1,4 @@
-// Copyright 2023 Proyectos y Sistemas de Mantenimiento SL (eProsima).
+// Copyright 2016 Proyectos y Sistemas de Mantenimiento SL (eProsima).
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -14,9 +14,9 @@
 
 /*!
  * @file DynamicTypesCollection.cpp
- * This source file contains the definition of the described types in the IDL file.
+ * This source file contains the implementation of the described types in the IDL file.
  *
- * This file was generated by the tool gen.
+ * This file was generated by the tool fastddsgen.
  */
 
 #ifdef _WIN32
@@ -26,34 +26,93 @@ char dummy;
 }  // namespace
 #endif  // _WIN32
 
-#include <ddsrecorder_participants/common/types/dynamic_types/v1/DynamicTypesCollection.hpp>
+#include <ddsrecorder_participants/common/types/dynamic_types_collection/v2/DynamicTypesCollection.hpp>
+
+#if FASTCDR_VERSION_MAJOR == 1
+
 #include <fastcdr/Cdr.h>
 
+
 #include <fastcdr/exceptions/BadParamException.h>
+using namespace eprosima::fastcdr::exception;
 
 #include <utility>
 
+namespace helper {
+namespace internal {
+
+enum class Size
+{
+    UInt8,
+    UInt16,
+    UInt32,
+    UInt64,
+};
+
+constexpr Size get_size(
+        int s)
+{
+    return (s <= 8 ) ? Size::UInt8:
+           (s <= 16) ? Size::UInt16:
+           (s <= 32) ? Size::UInt32: Size::UInt64;
+}
+
+template<Size s>
+struct FindTypeH;
+
+template<>
+struct FindTypeH<Size::UInt8>
+{
+    using type = std::uint8_t;
+};
+
+template<>
+struct FindTypeH<Size::UInt16>
+{
+    using type = std::uint16_t;
+};
+
+template<>
+struct FindTypeH<Size::UInt32>
+{
+    using type = std::uint32_t;
+};
+
+template<>
+struct FindTypeH<Size::UInt64>
+{
+    using type = std::uint64_t;
+};
+} // namespace internal
+
+template<int S>
+struct FindType
+{
+    using type = typename internal::FindTypeH<internal::get_size(S)>::type;
+};
+} // namespace helper
+
+#define eprosima_ddsrecorder_participants_DynamicTypesCollection_max_cdr_typesize 78412ULL;
+#define eprosima_ddsrecorder_participants_DynamicType_max_cdr_typesize 784ULL;
+
+
 namespace eprosima {
+
 namespace ddsrecorder {
+
 namespace participants {
 
-using namespace eprosima::fastcdr::exception;
 
-#define DynamicTypesCollection_max_cdr_typesize 78004ULL;
-#define DynamicType_max_cdr_typesize 780ULL;
-#define DynamicTypesCollection_max_key_cdr_typesize 0ULL;
-#define DynamicType_max_key_cdr_typesize 0ULL;
 
 DynamicType::DynamicType()
 {
-    // string m_type_name
-    m_type_name = "";
+    // /type_d() m_type_name
 
-    // string m_type_information
-    m_type_information = "";
+    // /type_d() m_type_information
 
-    // string m_type_object
-    m_type_object = "";
+    // /type_d() m_type_object
+
+
 }
 
 DynamicType::~DynamicType()
@@ -64,23 +123,37 @@ DynamicType::DynamicType(
         const DynamicType& x)
 {
     m_type_name = x.m_type_name;
+
+
     m_type_information = x.m_type_information;
+
+
     m_type_object = x.m_type_object;
+
 }
 
 DynamicType::DynamicType(
         DynamicType&& x) noexcept
 {
     m_type_name = std::move(x.m_type_name);
+
+
     m_type_information = std::move(x.m_type_information);
+
+
     m_type_object = std::move(x.m_type_object);
+
 }
 
 DynamicType& DynamicType::operator =(
         const DynamicType& x)
 {
     m_type_name = x.m_type_name;
+
+
     m_type_information = x.m_type_information;
+
+
     m_type_object = x.m_type_object;
 
     return *this;
@@ -90,7 +163,11 @@ DynamicType& DynamicType::operator =(
         DynamicType&& x) noexcept
 {
     m_type_name = std::move(x.m_type_name);
+
+
     m_type_information = std::move(x.m_type_information);
+
+
     m_type_object = std::move(x.m_type_object);
 
     return *this;
@@ -99,7 +176,8 @@ DynamicType& DynamicType::operator =(
 bool DynamicType::operator ==(
         const DynamicType& x) const
 {
-    return (m_type_name == x.m_type_name && m_type_information == x.m_type_information &&
+    return (m_type_name == x.m_type_name &&
+           m_type_information == x.m_type_information &&
            m_type_object == x.m_type_object);
 }
 
@@ -113,20 +191,25 @@ size_t DynamicType::getMaxCdrSerializedSize(
         size_t current_alignment)
 {
     static_cast<void>(current_alignment);
-    return DynamicType_max_cdr_typesize;
+    return eprosima_ddsrecorder_participants_DynamicType_max_cdr_typesize;
 }
 
 size_t DynamicType::getCdrSerializedSize(
         const DynamicType& data,
         size_t current_alignment)
 {
-    static_cast<void>(data);
+    (void)data;
     size_t initial_alignment = current_alignment;
 
     current_alignment += 4 + eprosima::fastcdr::Cdr::alignment(current_alignment, 4) + data.type_name().size() + 1;
+
+
     current_alignment += 4 +
             eprosima::fastcdr::Cdr::alignment(current_alignment, 4) + data.type_information().size() + 1;
+
+
     current_alignment += 4 + eprosima::fastcdr::Cdr::alignment(current_alignment, 4) + data.type_object().size() + 1;
+
 
     return current_alignment - initial_alignment;
 }
@@ -134,15 +217,39 @@ size_t DynamicType::getCdrSerializedSize(
 void DynamicType::serialize(
         eprosima::fastcdr::Cdr& scdr) const
 {
-    scdr << m_type_name.c_str(); scdr << m_type_information.c_str(); scdr << m_type_object.c_str();
+    scdr << m_type_name.c_str();
+
+    scdr << m_type_information.c_str();
+
+    scdr << m_type_object.c_str();
+
 }
 
 void DynamicType::deserialize(
         eprosima::fastcdr::Cdr& dcdr)
 {
     dcdr >> m_type_name;
+
+
+
     dcdr >> m_type_information;
+
+
+
     dcdr >> m_type_object;
+
+
+}
+
+bool DynamicType::isKeyDefined()
+{
+    return false;
+}
+
+void DynamicType::serializeKey(
+        eprosima::fastcdr::Cdr& scdr) const
+{
+    (void) scdr;
 }
 
 /*!
@@ -259,26 +366,11 @@ std::string& DynamicType::type_object()
     return m_type_object;
 }
 
-size_t DynamicType::getKeyMaxCdrSerializedSize(
-        size_t current_alignment)
-{
-    static_cast<void>(current_alignment);
-    return DynamicType_max_key_cdr_typesize;
-}
-
-bool DynamicType::isKeyDefined()
-{
-    return false;
-}
-
-void DynamicType::serializeKey(
-        eprosima::fastcdr::Cdr& scdr) const
-{
-    static_cast<void>(scdr);
-}
-
 DynamicTypesCollection::DynamicTypesCollection()
 {
+    // sequence<eprosima::ddsrecorder::participants::DynamicType> m_dynamic_types
+
+
 }
 
 DynamicTypesCollection::~DynamicTypesCollection()
@@ -289,12 +381,14 @@ DynamicTypesCollection::DynamicTypesCollection(
         const DynamicTypesCollection& x)
 {
     m_dynamic_types = x.m_dynamic_types;
+
 }
 
 DynamicTypesCollection::DynamicTypesCollection(
         DynamicTypesCollection&& x) noexcept
 {
     m_dynamic_types = std::move(x.m_dynamic_types);
+
 }
 
 DynamicTypesCollection& DynamicTypesCollection::operator =(
@@ -329,22 +423,26 @@ size_t DynamicTypesCollection::getMaxCdrSerializedSize(
         size_t current_alignment)
 {
     static_cast<void>(current_alignment);
-    return DynamicTypesCollection_max_cdr_typesize;
+    return eprosima_ddsrecorder_participants_DynamicTypesCollection_max_cdr_typesize;
 }
 
 size_t DynamicTypesCollection::getCdrSerializedSize(
         const DynamicTypesCollection& data,
         size_t current_alignment)
 {
-    static_cast<void>(data);
+    (void)data;
     size_t initial_alignment = current_alignment;
 
     current_alignment += 4 + eprosima::fastcdr::Cdr::alignment(current_alignment, 4);
 
+
     for (size_t a = 0; a < data.dynamic_types().size(); ++a)
     {
-        current_alignment += DynamicType::getCdrSerializedSize(data.dynamic_types().at(a), current_alignment);
+        current_alignment += eprosima::ddsrecorder::participants::DynamicType::getCdrSerializedSize(
+            data.dynamic_types().at(a), current_alignment);
     }
+
+
 
     return current_alignment - initial_alignment;
 }
@@ -353,57 +451,16 @@ void DynamicTypesCollection::serialize(
         eprosima::fastcdr::Cdr& scdr) const
 {
     scdr << m_dynamic_types;
+
+
 }
 
 void DynamicTypesCollection::deserialize(
         eprosima::fastcdr::Cdr& dcdr)
 {
     dcdr >> m_dynamic_types;
-}
 
-/*!
- * @brief This function copies the value in member dynamic_types
- * @param _dynamic_types New value to be copied in member dynamic_types
- */
-void DynamicTypesCollection::dynamic_types(
-        const std::vector<DynamicType>& _dynamic_types)
-{
-    m_dynamic_types = _dynamic_types;
-}
 
-/*!
- * @brief This function moves the value in member dynamic_types
- * @param _dynamic_types New value to be moved in member dynamic_types
- */
-void DynamicTypesCollection::dynamic_types(
-        std::vector<DynamicType>&& _dynamic_types)
-{
-    m_dynamic_types = std::move(_dynamic_types);
-}
-
-/*!
- * @brief This function returns a constant reference to member dynamic_types
- * @return Constant reference to member dynamic_types
- */
-const std::vector<DynamicType>& DynamicTypesCollection::dynamic_types() const
-{
-    return m_dynamic_types;
-}
-
-/*!
- * @brief This function returns a reference to member dynamic_types
- * @return Reference to member dynamic_types
- */
-std::vector<DynamicType>& DynamicTypesCollection::dynamic_types()
-{
-    return m_dynamic_types;
-}
-
-size_t DynamicTypesCollection::getKeyMaxCdrSerializedSize(
-        size_t current_alignment)
-{
-    static_cast<void>(current_alignment);
-    return DynamicTypesCollection_max_key_cdr_typesize;
 }
 
 bool DynamicTypesCollection::isKeyDefined()
@@ -414,9 +471,53 @@ bool DynamicTypesCollection::isKeyDefined()
 void DynamicTypesCollection::serializeKey(
         eprosima::fastcdr::Cdr& scdr) const
 {
-    static_cast<void>(scdr);
+    (void) scdr;
 }
 
-} /* namespace participants */
-} /* namespace ddsrecorder */
-} /* namespace eprosima */
+/*!
+ * @brief This function copies the value in member dynamic_types
+ * @param _dynamic_types New value to be copied in member dynamic_types
+ */
+void DynamicTypesCollection::dynamic_types(
+        const std::vector<eprosima::ddsrecorder::participants::DynamicType>& _dynamic_types)
+{
+    m_dynamic_types = _dynamic_types;
+}
+
+/*!
+ * @brief This function moves the value in member dynamic_types
+ * @param _dynamic_types New value to be moved in member dynamic_types
+ */
+void DynamicTypesCollection::dynamic_types(
+        std::vector<eprosima::ddsrecorder::participants::DynamicType>&& _dynamic_types)
+{
+    m_dynamic_types = std::move(_dynamic_types);
+}
+
+/*!
+ * @brief This function returns a constant reference to member dynamic_types
+ * @return Constant reference to member dynamic_types
+ */
+const std::vector<eprosima::ddsrecorder::participants::DynamicType>& DynamicTypesCollection::dynamic_types() const
+{
+    return m_dynamic_types;
+}
+
+/*!
+ * @brief This function returns a reference to member dynamic_types
+ * @return Reference to member dynamic_types
+ */
+std::vector<eprosima::ddsrecorder::participants::DynamicType>& DynamicTypesCollection::dynamic_types()
+{
+    return m_dynamic_types;
+}
+
+} // namespace participants
+
+
+} // namespace ddsrecorder
+
+
+} // namespace eprosima
+
+#endif // FASTCDR_VERSION_MAJOR == 1
